+++ httpx/api.py
@@ -6,6 +6,7 @@ from .models import (
     AuthTypes,
     CookieTypes,
     HeaderTypes,
+    ProxiesTypes,
     QueryParamTypes,
     RequestData,
     RequestFiles,
@@ -27,11 +28,11 @@ def request(
     auth: AuthTypes = None,
     timeout: TimeoutTypes = None,
     allow_redirects: bool = True,
-    # proxies
     cert: CertTypes = None,
     verify: VerifyTypes = True,
     stream: bool = False,
     trust_env: bool = None,
+    proxies: ProxiesTypes = None,
 ) -> Response:
     with Client(http_versions=["HTTP/1.1"]) as client:
         return client.request(
@@ -66,6 +67,7 @@ def get(
     verify: VerifyTypes = True,
     timeout: TimeoutTypes = None,
     trust_env: bool = None,
+    proxies: ProxiesTypes = None,
 ) -> Response:
     return request(
         "GET",
@@ -96,6 +98,7 @@ def options(
     verify: VerifyTypes = True,
     timeout: TimeoutTypes = None,
     trust_env: bool = None,
+    proxies: ProxiesTypes = None,
 ) -> Response:
     return request(
         "OPTIONS",
@@ -126,6 +129,7 @@ def head(
     verify: VerifyTypes = True,
     timeout: TimeoutTypes = None,
     trust_env: bool = None,
+    proxies: ProxiesTypes = None,
 ) -> Response:
     return request(
         "HEAD",
@@ -159,6 +163,7 @@ def post(
     verify: VerifyTypes = True,
     timeout: TimeoutTypes = None,
     trust_env: bool = None,
+    proxies: ProxiesTypes = None,
 ) -> Response:
     return request(
         "POST",
@@ -195,6 +200,7 @@ def put(
     verify: VerifyTypes = True,
     timeout: TimeoutTypes = None,
     trust_env: bool = None,
+    proxies: ProxiesTypes = None,
 ) -> Response:
     return request(
         "PUT",
@@ -231,6 +237,7 @@ def patch(
     verify: VerifyTypes = True,
     timeout: TimeoutTypes = None,
     trust_env: bool = None,
+    proxies: ProxiesTypes = None,
 ) -> Response:
     return request(
         "PATCH",
@@ -267,6 +274,7 @@ def delete(
     verify: VerifyTypes = True,
     timeout: TimeoutTypes = None,
     trust_env: bool = None,
+    proxies: ProxiesTypes = None,
 ) -> Response:
     return request(
         "DELETE",