+++ httpx/config.py
@@ -6,6 +6,7 @@ from pathlib import Path
 import certifi
 
 from .__version__ import __version__
+from .utils import get_ca_bundle_from_env, get_logger
 
 CertTypes = typing.Union[str, typing.Tuple[str, str], typing.Tuple[str, str, str]]
 VerifyTypes = typing.Union[str, bool, ssl.SSLContext]
@@ -37,6 +38,9 @@ DEFAULT_CIPHERS = ":".join(
         "!DSS",
     ]
 )
+
+
+logger = get_logger(__name__)
 
 
 class SSLConfig:
@@ -90,6 +94,14 @@ class SSLConfig:
     ) -> ssl.SSLContext:
         http_versions = HTTPVersionConfig() if http_versions is None else http_versions
 
+        logger.debug(
+            f"load_ssl_context "
+            f"verify={self.verify!r} "
+            f"cert={self.cert!r} "
+            f"trust_env={self.trust_env!r} "
+            f"http_versions={http_versions!r}"
+        )
+
         if self.ssl_context is None:
             self.ssl_context = (
                 self.load_ssl_context_verify(http_versions=http_versions)
@@ -117,6 +129,11 @@ class SSLConfig:
         """
         Return an SSL context for verified connections.
         """
+        if self.trust_env and self.verify is True:
+            ca_bundle = get_ca_bundle_from_env()
+            if ca_bundle is not None:
+                self.verify = ca_bundle  # type: ignore
+
         if isinstance(self.verify, bool):
             ca_bundle_path = DEFAULT_CA_BUNDLE_PATH
         elif Path(self.verify).exists():
@@ -146,8 +163,10 @@ class SSLConfig:
             pass
 
         if ca_bundle_path.is_file():
+            logger.debug(f"load_verify_locations cafile={ca_bundle_path!s}")
             context.load_verify_locations(cafile=str(ca_bundle_path))
         elif ca_bundle_path.is_dir():
+            logger.debug(f"load_verify_locations capath={ca_bundle_path!s}")
             context.load_verify_locations(capath=str(ca_bundle_path))
 
         self._load_client_certs(context)
@@ -269,8 +288,8 @@ class HTTPVersionConfig:
             }
         else:
             raise TypeError(
-                f"HTTP version should be a string or list of strings, "
-                "but got {type(http_versions)}"
+                "HTTP version should be a string or list of strings, "
+                f"but got {type(http_versions)}"
             )
 
         for version in self.http_versions:
@@ -278,7 +297,7 @@ class HTTPVersionConfig:
                 raise ValueError(f"Unsupported HTTP version {version!r}.")
 
         if not self.http_versions:
-            raise ValueError(f"HTTP versions cannot be an empty list.")
+            raise ValueError("HTTP versions cannot be an empty list.")
 
     @property
     def alpn_identifiers(self) -> typing.List[str]: