+++ httpx/utils.py
@@ -5,7 +5,11 @@ import os
 import re
 import sys
 import typing
+from datetime import timedelta
 from pathlib import Path
+from time import perf_counter
+from types import TracebackType
+from urllib.request import getproxies
 
 
 def normalize_header_key(value: typing.AnyStr, encoding: str = None) -> bytes:
@@ -107,6 +111,18 @@ def get_netrc_login(host: str) -> typing.Optional[typing.Tuple[str, str, str]]:
     return netrc_info.authenticators(host)  # type: ignore
 
 
+def get_ca_bundle_from_env() -> typing.Optional[str]:
+    if "SSL_CERT_FILE" in os.environ:
+        ssl_file = Path(os.environ["SSL_CERT_FILE"])
+        if ssl_file.is_file():
+            return str(ssl_file)
+    if "SSL_CERT_DIR" in os.environ:
+        ssl_path = Path(os.environ["SSL_CERT_DIR"])
+        if ssl_path.is_dir():
+            return str(ssl_path)
+    return None
+
+
 def parse_header_links(value: str) -> typing.List[typing.Dict[str, str]]:
     """
     Returns a list of parsed link headers, for more info see:
@@ -144,6 +160,18 @@ def parse_header_links(value: str) -> typing.List[typing.Dict[str, str]]:
     return links
 
 
+SENSITIVE_HEADERS = {"authorization", "proxy-authorization"}
+
+
+def obfuscate_sensitive_headers(
+    items: typing.Iterable[typing.Tuple[typing.AnyStr, typing.AnyStr]]
+) -> typing.Iterator[typing.Tuple[typing.AnyStr, typing.AnyStr]]:
+    for k, v in items:
+        if to_str(k.lower()) in SENSITIVE_HEADERS:
+            v = to_bytes_or_str("[secure]", match_type_of=v)
+        yield k, v
+
+
 _LOGGER_INITIALIZED = False
 
 
@@ -171,15 +199,143 @@ def get_logger(name: str) -> logging.Logger:
     return logging.getLogger(name)
 
 
+def kv_format(**kwargs: typing.Any) -> str:
+    """Format arguments into a key=value line.
+
+    >>> formatkv(x=1, name="Bob")
+    "x=1 name='Bob'"
+    """
+    return " ".join(f"{key}={value!r}" for key, value in kwargs.items())
+
+
+def get_environment_proxies() -> typing.Dict[str, str]:
+    """Gets proxy information from the environment"""
+
+    # urllib.request.getproxies() falls back on System
+    # Registry and Config for proxies on Windows and macOS.
+    # We don't want to propagate non-HTTP proxies into
+    # our configuration such as 'TRAVIS_APT_PROXY'.
+    proxies = {
+        key: val
+        for key, val in getproxies().items()
+        if ("://" in key or key in ("http", "https"))
+    }
+
+    # Favor lowercase environment variables over uppercase.
+    all_proxy = get_environ_lower_and_upper("ALL_PROXY")
+    if all_proxy is not None:
+        proxies["all"] = all_proxy
+
+    return proxies
+
+
+def get_environ_lower_and_upper(key: str) -> typing.Optional[str]:
+    """Gets a value from os.environ with both the lowercase and uppercase
+    environment variable. Prioritizes the lowercase environment variable.
+    """
+    for key in (key.lower(), key.upper()):
+        value = os.environ.get(key, None)
+        if value is not None and isinstance(value, str):
+            return value
+    return None
+
+
 def to_bytes(value: typing.Union[str, bytes], encoding: str = "utf-8") -> bytes:
     return value.encode(encoding) if isinstance(value, str) else value
 
 
-def to_str(str_or_bytes: typing.Union[str, bytes], encoding: str = "utf-8") -> str:
-    return (
-        str_or_bytes if isinstance(str_or_bytes, str) else str_or_bytes.decode(encoding)
-    )
+def to_str(value: typing.Union[str, bytes], encoding: str = "utf-8") -> str:
+    return value if isinstance(value, str) else value.decode(encoding)
+
+
+def to_bytes_or_str(value: str, match_type_of: typing.AnyStr) -> typing.AnyStr:
+    return value if isinstance(match_type_of, str) else value.encode()
 
 
 def unquote(value: str) -> str:
     return value[1:-1] if value[0] == value[-1] == '"' else value
+
+
+class ElapsedTimer:
+    def __init__(self) -> None:
+        self.start: float = perf_counter()
+        self.end: typing.Optional[float] = None
+
+    def __enter__(self) -> "ElapsedTimer":
+        self.start = perf_counter()
+        return self
+
+    def __exit__(
+        self,
+        exc_type: typing.Type[BaseException] = None,
+        exc_value: BaseException = None,
+        traceback: TracebackType = None,
+    ) -> None:
+        self.end = perf_counter()
+
+    @property
+    def elapsed(self) -> timedelta:
+        if self.end is None:
+            return timedelta(seconds=perf_counter() - self.start)
+        return timedelta(seconds=self.end - self.start)
+
+
+ASGI_PLACEHOLDER_FORMAT = {
+    "body": "<{length} bytes>",
+    "bytes": "<{length} bytes>",
+    "text": "<{length} chars>",
+}
+
+
+def asgi_message_with_placeholders(message: dict) -> dict:
+    """
+    Return an ASGI message, with any body-type content omitted and replaced
+    with a placeholder.
+    """
+    new_message = message.copy()
+
+    for attr in ASGI_PLACEHOLDER_FORMAT:
+        if attr in message:
+            content = message[attr]
+            placeholder = ASGI_PLACEHOLDER_FORMAT[attr].format(length=len(content))
+            new_message[attr] = placeholder
+
+    if "headers" in message:
+        new_message["headers"] = list(obfuscate_sensitive_headers(message["headers"]))
+
+    return new_message
+
+
+class MessageLoggerASGIMiddleware:
+    def __init__(self, app: typing.Callable, logger: logging.Logger) -> None:
+        self.app = app
+        self.logger = logger
+
+    async def __call__(
+        self, scope: dict, receive: typing.Callable, send: typing.Callable
+    ) -> None:
+        async def inner_receive() -> dict:
+            message = await receive()
+            logged_message = asgi_message_with_placeholders(message)
+            self.logger.debug(f"sent {kv_format(**logged_message)}")
+            return message
+
+        async def inner_send(message: dict) -> None:
+            logged_message = asgi_message_with_placeholders(message)
+            self.logger.debug(f"received {kv_format(**logged_message)}")
+            await send(message)
+
+        logged_scope = dict(scope)
+        if "headers" in scope:
+            logged_scope["headers"] = list(
+                obfuscate_sensitive_headers(scope["headers"])
+            )
+        self.logger.debug(f"started {kv_format(**logged_scope)}")
+
+        try:
+            await self.app(scope, inner_receive, inner_send)
+        except BaseException as exc:
+            self.logger.debug("raised_exception")
+            raise exc from None
+        else:
+            self.logger.debug("completed")