+++ tests/dispatch/test_connections.py
@@ -15,12 +15,12 @@ async def test_post(server, backend):
         assert response.status_code == 200
 
 
-async def test_https_get_with_ssl_defaults(https_server, backend):
+async def test_https_get_with_ssl_defaults(https_server, ca_cert_pem_file, backend):
     """
     An HTTPS request, with default SSL configuration set on the client.
     """
     async with HTTPConnection(
-        origin=https_server.url, verify=False, backend=backend
+        origin=https_server.url, verify=ca_cert_pem_file, backend=backend
     ) as conn:
         response = await conn.request("GET", https_server.url)
         await response.read()
@@ -28,12 +28,12 @@ async def test_https_get_with_ssl_defaults(https_server, backend):
         assert response.content == b"Hello, world!"
 
 
-async def test_https_get_with_sll_overrides(https_server, backend):
+async def test_https_get_with_sll_overrides(https_server, ca_cert_pem_file, backend):
     """
     An HTTPS request, with SSL configuration set on the request.
     """
     async with HTTPConnection(origin=https_server.url, backend=backend) as conn:
-        response = await conn.request("GET", https_server.url, verify=False)
+        response = await conn.request("GET", https_server.url, verify=ca_cert_pem_file)
         await response.read()
         assert response.status_code == 200
         assert response.content == b"Hello, world!"